+++ cfgov/cfgov/settings/base.py
@@ -34,6 +34,7 @@ INSTALLED_APPS = (
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'storages',
     'localflavor',
     'flags',
     'v1',
@@ -80,6 +81,11 @@ TEMPLATES = [
         'APP_DIRS': False,
         'OPTIONS': {
             'environment': 'v1.environment',
+            'extensions': [
+                'wagtail.wagtailcore.templatetags.jinja2.core',
+                'wagtail.wagtailadmin.templatetags.jinja2.userbar',
+                'wagtail.wagtailimages.templatetags.jinja2.images',
+            ],
         }
     },
 ]
@@ -140,7 +146,7 @@ STATICFILES_STORAGE = 'django.contrib.staticfiles.storage.StaticFilesStorage'
 # app-specific 'static' directories.
 STATICFILES_DIRS = [
     PROJECT_ROOT.child('static_built'),
-    ('legacy', PROJECT_ROOT.child('v1','static-legacy')),
+    ('legacy', PROJECT_ROOT.child('v1', 'static-legacy')),
 ]
 
 ALLOWED_HOSTS = ['*']
@@ -255,3 +261,15 @@ SHEER_ELASTICSEARCH_SETTINGS = \
 
 # PDFReactor
 PDFREACTOR_LIB = os.environ.get('PDFREACTOR_LIB', '/opt/PDFreactor/wrappers/python/lib')
+
+# S3 Configuration
+if os.environ.get('S3_ENABLED', False):
+    DEFAULT_FILE_STORAGE = 'v1.s3utils.MediaRootS3BotoStorage'
+    AWS_S3_SECURE_URLS = False  # True = use https; False = use http
+    AWS_QUERYSTRING_AUTH = False  # False = do not use authentication-related query parameters for requests
+    AWS_S3_ACCESS_KEY_ID = os.environ.get('AWS_S3_ACCESS_KEY_ID')
+    AWS_S3_SECRET_ACCESS_KEY = os.environ.get('AWS_S3_SECRET_ACCESS_KEY')
+    AWS_STORAGE_BUCKET_NAME = os.environ.get('AWS_STORAGE_BUCKET_NAME')
+    AWS_S3_CALLING_FORMAT = 'boto.s3.connection.OrdinaryCallingFormat'
+
+    MEDIA_URL = os.environ.get('AWS_S3_URL') + '/f/'