+++ cfgov/v1/admin.py
@@ -1,3 +1,7 @@
 from django.contrib import admin
+from models.snippets import Contact
 
-# Register your models here.
+@admin.register(Contact)
+class ContactAdmin(admin.ModelAdmin):
+    pass
+