+++ cfgov/v1/models/demo.py
@@ -5,34 +5,53 @@ from wagtail.wagtailadmin.edit_handlers import FieldPanel, MultiFieldPanel, \
     FieldRowPanel, TabbedInterface, ObjectList, StreamFieldPanel
 from wagtail.wagtailimages.edit_handlers import ImageChooserPanel
 from wagtail.wagtaildocs.edit_handlers import DocumentChooserPanel
+from wagtail.wagtailsnippets.edit_handlers import SnippetChooserPanel
 from wagtail.wagtailcore import blocks
 
 from .base import CFGOVPage
-from .molecules import HalfWidthLinkBlob, TextIntroduction, ImageText5050
-from .organisms import Well
+from . import atoms
+from . import molecules
+from . import organisms
+from .snippets import Contact
 
 
 class DemoPage(CFGOVPage):
     molecules = StreamField([
-        ('half_width_link_blob', HalfWidthLinkBlob()),
-        ('text_introduction', TextIntroduction()),
-        ('image_text_5050', ImageText5050())
+        ('half_width_link_blob', molecules.HalfWidthLinkBlob()),
+        ('text_introduction', molecules.TextIntroduction()),
+        ('image_text_2575', molecules.ImageText2575()),
+        ('image_text_5050', molecules.ImageText5050()),
+        ('hero', molecules.Hero()),
+        ('formfield_with_button', molecules.FormFieldWithButton()),
+        ('call_to_action', molecules.CallToAction())
     ], blank=True)
 
     organisms = StreamField([
-        ('well', Well())
+        ('well', organisms.Well()),
+        ('full_width_text', organisms.FullWidthText()),
+        ('post_preview', organisms.PostPreview()),
     ], blank=True)
+
+    contact = models.ForeignKey(
+        Contact,
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+'
+    )
+
 
     # General content tab
     content_panels = CFGOVPage.content_panels + [
         StreamFieldPanel('molecules'),
         StreamFieldPanel('organisms'),
+        SnippetChooserPanel('contact'),
     ]
 
     # Tab handler interface
     edit_handler = TabbedInterface([
         ObjectList(content_panels, heading='General Content'),
-        ObjectList(CFGOVPage.promote_panels, heading='Promote'),
+        ObjectList(CFGOVPage.sidefoot_panels, heading='Promote'),
         ObjectList(CFGOVPage.settings_panels, heading='Settings', classname="settings"),
     ])
 
@@ -41,4 +60,4 @@ class DemoPage(CFGOVPage):
         return context
 
     def get_template(self, request):
-        return 'v1/demo/index.html'
+        return 'wagtail/demo/index.html'