+++ cfgov/v1/models/ref.py
@@ -1,0 +1,15 @@
+from wagtail.wagtailcore import blocks
+
+choices = [
+    ('info for consumers', 'Info for consumers'),
+    ('at the cfpb', 'At the CFPB'),
+    ('data, research & reports', 'Data, research & reports'),
+    ('policy & compliance', 'Policy & Compliance'),
+    ('speech', 'Speech'),
+    ('press release', 'Press Release'),
+    ('op-ed', 'Op-ed'),
+    ('testimony', 'Testimony'),
+    ('cfpb_report', 'CFPB Report'),
+    ('blog', 'Blog'),
+    ('newsroom', 'Newsroom'),
+]