+++ test/unit_tests/modules/footer-button-spec.js
@@ -12,7 +12,8 @@ describe( 'The Footer Return-to-Top Button', function() {
 
   before( function() {
     $ = require( 'jquery' );
-    footerButton = require( '../../../cfgov/preprocessed/js/modules/footer-button.js' );
+    footerButton =
+      require( '../../../cfgov/unprocessed/js/modules/footer-button.js' );
     sandbox = sinon.sandbox.create();
   } );